feat: baby steps towards session management
I removed the GraphQL things for now as the basis for that is a working
connection and on top of that session management. This is the first step
towards that.

Currently the frontend connects without a session token, which would be
equivalent to a new visitor. The server responds with an anonymous
session token which the client has to store for later use. On subsequent
page loads (refresh, new tab) the client has to send her token along
when attempting to establish a websocket connection. The server will
validate it and it will either be valid or not.

If it is valid, the connection (socket) will be assigned to the session
that has been started prior. If not, a new session token will be
generated and sent back to the client. From here on the cycle repeats
itself.

So far this has almost nothing to do with being logged in or not.
"Logging in" basically means "upgrading" an anonymous session to map to
a real user.